requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-huggingface
huggingface_hub[inference]
sentence-transformers
faiss-cpu
pypdf
transformers